travis ci build fix attempt #4

+++ src/internal/kvr_msgpack.cpp
@@ -440,7 +440,7 @@ public:
           uint16_t len = 0;
           success = this->get ((uint8_t *) &len, 2);
           uint16_t slen = bigendian16 (len);
-          KVR_ASSERT (slen <= kvr::SZ_T_MAX);
+          KVR_ASSERT ((uint64_t) slen <= kvr::SZ_T_MAX);
           const char *str = success ? (const char *) this->push (slen) : NULL;
           success = str ? ctx.read_string (str, slen) : false;
           break;
@@ -451,7 +451,7 @@ public:
           uint32_t len = 0;
           success = this->get ((uint8_t *) &len, 4);
           uint32_t slen = bigendian32 (len);
-          KVR_ASSERT (slen <= kvr::SZ_T_MAX);
+          KVR_ASSERT ((uint64_t) slen <= kvr::SZ_T_MAX);
           const char *str = success ? (const char *) this->push (slen) : NULL;
           success = str ? ctx.read_string (str, slen) : false;
           break;
@@ -462,7 +462,7 @@ public:
           uint16_t len = 0;
           success = this->get ((uint8_t *) &len, 2);
           uint16_t alen = bigendian16 (len);
-          KVR_ASSERT (alen <= kvr::SZ_T_MAX);
+          KVR_ASSERT ((uint64_t) alen <= kvr::SZ_T_MAX);
           bool ok = success && ctx.read_array_start (alen);
           for (uint16_t i = 0; ok && (i < alen); ++i)
           {
@@ -480,7 +480,7 @@ public:
           uint32_t len = 0;
           success = this->get ((uint8_t *) &len, 4);
           uint32_t alen = bigendian32 (len);
-          KVR_ASSERT (alen <= kvr::SZ_T_MAX);
+          KVR_ASSERT ((uint64_t) alen <= kvr::SZ_T_MAX);
           bool ok = success && ctx.read_array_start (alen);
           for (uint32_t i = 0; ok && (i < alen); ++i)
           {
@@ -499,7 +499,7 @@ public:
           uint16_t len = 0;
           success = this->get ((uint8_t *) &len, 2);
           uint16_t msz = bigendian16 (len);
-          KVR_ASSERT (msz <= kvr::SZ_T_MAX);
+          KVR_ASSERT ((uint64_t) msz <= kvr::SZ_T_MAX);
           bool ok = success && ctx.read_map_start (msz);
           for (uint16_t i = 0; ok && (i < msz); ++i)
           {
@@ -518,7 +518,7 @@ public:
           uint32_t len = 0;
           success = this->get ((uint8_t *) &len, 4);
           uint32_t msz = bigendian32 (len);
-          KVR_ASSERT (msz <= kvr::SZ_T_MAX);
+          KVR_ASSERT ((uint64_t) msz <= kvr::SZ_T_MAX);
           bool ok = success && ctx.read_map_start (msz);
           for (uint32_t i = 0; ok && (i < msz); ++i)
           {
@@ -701,7 +701,7 @@ private:
           uint16_t len = 0;
           success = this->get ((uint8_t *) &len, 2);
           uint16_t slen = bigendian16 (len);
-          KVR_ASSERT (slen <= kvr::SZ_T_MAX);
+          KVR_ASSERT ((uint64_t) slen <= kvr::SZ_T_MAX);
           const char *str = success ? (const char *) this->push (slen) : NULL;
           success = str ? ctx.read_key (str, slen) : false;
           break;
@@ -712,7 +712,7 @@ private:
           uint32_t len = 0;
           success = this->get ((uint8_t *) &len, 4);
           uint32_t slen = bigendian32 (len);
-          KVR_ASSERT (slen <= kvr::SZ_T_MAX);
+          KVR_ASSERT ((uint64_t) slen <= kvr::SZ_T_MAX);
           const char *str = success ? (const char *) this->push (slen) : NULL;
           success = str ? ctx.read_key (str, slen) : false;
           break;
